бекпап request

+++ catalog/controller/tool/upload.php
@@ -60,19 +60,39 @@ class ControllerToolUpload extends Controller {
 		}
 
 		if (!$json) {
-			$file = $filename . '.' . token(32);
-
-			move_uploaded_file($this->request->files['file']['tmp_name'], DIR_UPLOAD . $file);
-
-			// Hide the uploaded file name so people can not link to it directly.
-			$this->load->model('tool/upload');
-
-			$json['code'] = $this->model_tool_upload->addUpload($filename, $file);
-
-			$json['success'] = $this->language->get('text_upload');
+			$file = $filename . '.' . $this->request->post['size'];
+			
+			if(!file_exists(DIR_UPLOAD . $file)){
+				move_uploaded_file($this->request->files['file']['tmp_name'], DIR_UPLOAD . $file);
+	
+				// Hide the uploaded file name so people can not link to it directly.
+				$this->load->model('tool/upload');
+	
+				$json['code'] = $this->model_tool_upload->addUpload($filename, $file);
+				
+				$json['namefile'] = substr($filename, 0, 10);
+	
+				$json['success'] = $this->language->get('text_upload');
+			}else{
+				$json['error'] = $this->language->get('error_file');
+			}
 		}
 
 		$this->response->addHeader('Content-Type: application/json');
 		$this->response->setOutput(json_encode($json));
+	}
+	
+	public function remove(){
+		if($this->request->post['name'] && $this->request->post['size']){
+			$this->load->model('tool/upload');
+			
+			$name = $this->request->post['name'] . '.' . $this->request->post['size'];
+			
+			$normal = basename(preg_replace('/[^a-zA-Z0-9\.\-\s+]/', '', html_entity_decode($name, ENT_QUOTES, 'UTF-8')));
+			
+			unlink( DIR_UPLOAD . $normal);
+
+			$this->model_tool_upload->removeFile($normal);
+		}
 	}
 }
